requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nulltype==2.3.1
plaid-python==26.0.0
python-dateutil==2.9.0.post0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
